Move get_*_aspects function to reuse them for handling list requests
